More stuff

+++ src/main/scala/Challenges/OrderExercises.scala
@@ -10,17 +10,16 @@ import DS.Arrays.{Data, SubArraySumSolver}
  */
 object OrderExercises {
 
-  def arrayOrdering(array: Array[Int], k: Int): List[Int] = {
+  def arrayOrdering(array: Array[Long], k: Int): List[Long] = {
 
     def merge(l: List[Data], r: List[Data]): List[Data] = (l, r) match {
       case (Nil, Nil) => Nil
       case (xs, Nil) => xs
       case (Nil, ys) => ys
       case (x :: xs, y :: ys) => if (x > y) x :: merge(xs, r) else y :: merge(l, ys)
-
     }
 
-    def arrayOrderingTR(array: Array[Int]): List[Data] = {
+    def arrayOrderingTR(array: Array[Long]): List[Data] = {
       val solver = SubArraySumSolver(array)
       val data = solver.largestSubArraySum
       if (data.sum > 0) {
